Add support for reading device info via dinput
This is a WIP state.

SetupAPI is mainly useful for accessing keyboard and mouse. This is irrelevant for now since we want to focus on controller input instead.

After a lot of fiddling around with the changed Java foreign API in version 19 we finally achieved accessing the device information via a callback from the native Windows system assembly dinput8.dll.

This was particularly hard, since almost any reference you find on the internet was written in previous java versions, using the java.incubator.foreign API and the generated code by the jextract tool doesn't provide any usage examples either.