Impement support for foreground/exclusive cooperative level
This is required for some controllers to even work (like the Xbox 360 controllers) and yields.

Handle the case where devices are lost or not properly acquired by re-acquiring before polling. If we're operating in foreground/exclusive mode, losing window focus will yield in the device being lost. This still needs to be replaced by a proper "hot-swap" thread/solution.